in app notifications done

+++ lib/helpers/delete_venture.dart
@@ -4,13 +4,18 @@ import 'package:jet_palz/components/my_snack_bar.dart';
 
 Future<void> deleteVenture(
     BuildContext context, DocumentReference ventureRef) async {
+  DocumentReference? chatRef;
+
   final querySnapshot = await FirebaseFirestore.instance
       .collection("venture_chats")
       .where("venture", isEqualTo: ventureRef)
       .limit(1)
       .get();
 
-  final chatRef = querySnapshot.docs.first.reference;
+  if (querySnapshot.docs.isNotEmpty) {
+    chatRef = querySnapshot.docs.first.reference;
+  }
+
   final shouldDelete = await showDialog<bool>(
     context: context,
     builder: (BuildContext context) {
@@ -38,7 +43,22 @@ Future<void> deleteVenture(
   if (shouldDelete == true) {
     try {
       await ventureRef.delete();
-      await chatRef.delete();
+      if (chatRef != null) {
+        await chatRef.delete();
+      }
+      final requestsQuerySnapshot = await FirebaseFirestore.instance
+          .collection("requests")
+          .where("ventureId", isEqualTo: ventureRef.id)
+          .get();
+
+      // If there are any requests, delete them in a batch
+      if (requestsQuerySnapshot.docs.isNotEmpty) {
+        final batch = FirebaseFirestore.instance.batch();
+        for (final requestDoc in requestsQuerySnapshot.docs) {
+          batch.delete(requestDoc.reference);
+        }
+        await batch.commit();
+      }
       MySnackBar.show(context, content: Text('Venture deleted successfully'));
     } catch (e) {
       MySnackBar.show(context, content: Text('Failed to delete venture: $e'));
